+++ BSCapp/function.py
@@ -9,7 +9,36 @@ from django.db import connection
 from BSCapp.root_chain.utils import *
 from BSCapp.models import *
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from BSCapp.Paginator import *
+
+class JuncheePaginator(Paginator):
+    def __init__(self, object_list, per_page, range_num=5, orphans=0, allow_empty_first_page=True):
+        Paginator.__init__(self, object_list, per_page, orphans, allow_empty_first_page)
+        self.range_num = range_num
+
+    def page(self, number):
+        self.page_num = (int)(number)
+        return super(JuncheePaginator, self).page(number)
+
+    def _page_range_ext(self):
+        num_count = 2 * self.range_num + 1
+        if self.num_pages <= num_count:
+            return range(1, self.num_pages + 1)
+        num_list = []
+        num_list.append(self.page_num)
+        for i in range(1, self.range_num + 1):
+            if self.page_num - i <= 0:
+                num_list.append(num_count + self.page_num - i)
+            else:
+                num_list.append(self.page_num - i)
+
+            if self.page_num + i <= self.num_pages:
+                num_list.append(self.page_num + i)
+            else:
+                num_list.append(self.page_num + i - num_count)
+        num_list.sort()
+        return num_list
+
+    page_range_ext = property(_page_range_ext)
 
 def generate_sort_sql(table_name, sort_name, sort_type):
     sort_sql = 'order by ' + table_name + '.' + sort_name+' '+sort_type+';'
@@ -327,7 +356,7 @@ def noticeData_sql(user_id, sort_sql):
     return notices, unread_notices, unread_number
 
 
-def pagingData(request, datas, each_num):
+def pagingData(request, datas, each_num=10):
     # paginator = Paginator(datas, each_num)
     # print(each_num)
     paginator = JuncheePaginator(datas, each_num)
